Fixed problems till from 150 to 203 (#26)

+++ LeetCodeProblems/Stack/EvaluateReversePolishNotation.cpp
@@ -1,17 +1,15 @@
-#include <iostream>
 #include <stack>
-#include <vector>
 using namespace std;
 
 class Solution {
 public:
     int evalRPN(vector<string>& tokens) {
-        stack<int> stack;
+        stack<long long> stack;
         for(auto i: tokens){
             if(i == "+" || i == "*" || i == "/" || i == "-"){
-                int e1 = stack.top();
+                long long e1 = stack.top();
                 stack.pop();
-                int e2 = stack.top();
+                long long e2 = stack.top();
                 stack.pop();
                 if(i == "+"){
                     stack.push(e1+e2);
